Add pre/post up/down support for rooted GoBackend
When using the GoBackend on a rooted device allow for pre/post up/down
actions to be executed when the tunnel state changes. On non-rooted
devices the scripts are not executed but will still be parsed from the
configuration file. %i syntax is not supported. If any script fails to
execute the remaining scripts in that step are skipped

Signed-off-by: Adam Irr <adam.irr@outlook.com>